Add the PSF CoC to our own CoC.
As a Python community we should adhere to python standards.
With this in mind we are extending our CoC to encompass the official
PSF CoC found here: https://www.python.org/psf/codeofconduct/

We have made clear that this is the case and have added some
paraphrased parts of the PSF CoC to our pledge to emphasise the point.