Fix local development environment (#550)
* Updated docker-compose.yml file
* Added healthcheck to postgres container to check
   if it's ready to accept connections
* Updated postgresql port from 5432 to 5435 because
   the default port may already be in use by local
  postgres
* Removed version parameter from docker-compose.yml due
  to Docker emitting a message, "the attribute `version` is
  obsolete, it will be ignored, please remove it to avoid
  potential confusion"

* Upgrade pip as part of Dockerfile
* Upgrade pyyaml from 5.4.1 to 6.0.2

When first attempting to build the docker image, I received
an error when installing python dependencies. Upon further investigation,
the culprit was pyyaml wheel failing to build and the
fix was to upgrade pip and upgrade the pyyaml package.

* Upgrade aiohttp from 3.7.4.post0 to 3.10.11

After successfully building the docker image and running the
docker-compose build command, as per the CONTRIBUTING.md file, I
received the following error: "ValueError: Host 'localhost:8000' cannot
contain ':' (at position 9)".

Upon further investigation, the fix was to upgrade to version 3.10.11.